refactored knefile.js, fixed bug in db migration and seed process
