+++ knex/seeds/departmentSeed.js
@@ -1,10 +1,10 @@
 exports.seed = function (knex, Promise) {
   // Deletes ALL existing entries
-  return knex("employee")
+  return knex("departments")
     .del()
     .then(function () {
       // Inserts seed entries
-      return knex("employee").insert([
+      return knex("departments").insert([
         { 
             departmentName: "electronics" 
         },
