+++ knex/seeds/positionSeed.js
@@ -7,33 +7,33 @@ exports.seed = function (knex, Promise) {
       return knex("position").insert([
         {
           title: "Store associate",
-          department: "electronics",
           description: "assists the elecronics cumstomers",
           salary: "$45,000",
+          department_id: 1,
         },
         {
           title: "Cashier",
-          department: "automotive",
           description: "assists the automotive cumstomers",
           salary: "$45,000",
+          department_id: 2,
         },
         {
           title: "Cashier",
-          department: "sporting goods",
           description: "Checks out the customer",
           salary: "$24,000",
+          department_id: 3,
         },
         {
           title: "Store associate",
-          department: "automotive",
           description: "Handles the customer returns",
           salary: "$31,000",
+          department_id: 4,
         },
         {
           title: "Store associate",
-          department: "electronics",
           description: "Stocks shevles and assists customers",
           salary: "$21,000",
+          department_id: 5,
         },
       ]);
     });
