+++ knexfile.js
@@ -1,0 +1,24 @@
+// Create connection
+require('dotenv').config();
+
+module.exports = {
+  development: {
+    client: "mysql",
+    connection: {
+      host: process.env.DB_HOST,
+      port: process.env.DB_PORT,
+      user: process.env.DB_USER,
+      password: process.env.DB_PASSWORD,
+      database: process.env.DB_NAME,
+    },
+    migrations: {
+      directory: process.cwd() + '/knex/migrations',
+    },
+    seeds: {
+      directory: process.cwd() + '/knex/seeds',
+    },
+  },
+};
+
+
+
